requirements.txt
requests
beautifulsoup4
python-dotenv
langchain
chromadb
tiktoken
openai
panel
